Working on expectation calculation

+++ src/cs363/assignment2/Data.java
@@ -9,13 +9,13 @@ import java.util.Vector;
 public class Data {
 	
 	public String filename;
-	public static Vector<Vector<Integer>> column;
+	public static Vector<Vector<Double>> column;
 	
 	public Data(Path filePath, int columns){
 		
 		this.filename = filePath.toString();
-		column = new Vector<Vector<Integer>>(columns);
-		for(int i=0; i<3; i++) column.add(i, new Vector<Integer>(0));
+		column = new Vector<Vector<Double>>(columns);
+		for(int i=0; i<3; i++) column.add(i, new Vector<Double>(0));
 		
 		try { importData(filename); } 
 		catch (FileNotFoundException e) {}
@@ -26,8 +26,8 @@ public class Data {
 	public Data(String filename, int columns){
 		
 		this.filename = filename;
-		column = new Vector<Vector<Integer>>(columns);
-		for(int i=0; i<3; i++) column.add(i, new Vector<Integer>(0));
+		column = new Vector<Vector<Double>>(columns);
+		for(int i=0; i<3; i++) column.add(i, new Vector<Double>(0));
 		
 		try { importData(filename); } 
 		catch (FileNotFoundException e) {}
@@ -52,9 +52,8 @@ public class Data {
 		while(inFile.hasNext()){
 			
 			nextToken = inFile.next();
-			if(nextToken.equals("-")) column.get(i%3).add(-1);
-			//next = inFile.nextInt();
-			else column.get(i%3).add(Integer.parseInt(nextToken));
+			if(nextToken.equals("-")) column.get(i%3).add(-1.0);
+			else column.get(i%3).add(Double.parseDouble(nextToken));
 			i++; j++;
 		}
 	}
